+++ src/example_systems/continuous/RosslerLorenzUnidir6.jl
@@ -1,0 +1,68 @@
+using DynamicalSystemsBase: ContinuousDynamicalSystem
+using StaticArrays: SVector
+
+export RosslerLorenzUnidir6
+
+"""
+
+    RosslerLorenzUnidir6 <: ContinuousDefinition
+    RosslerLorenzUnidir6(; xi = [0.1, 0.2, 0.3, 0.05, 0.1, 0.15],
+        a₁ = 6, a₂ = 6, a₃ = 2.0, b₁ = 10, b₂ = 28, b₃ = 8/3, c_xy = 1.0)
+
+Initialise a Rössler-Lorenz system consisting of two independent 3D subsystems:
+one Rössler system and one Lorenz system. They are coupled such that the
+second component (`x₂`) of the Rössler system unidirectionally forces the
+second component (`y₂`) of the Lorenz system.
+
+The parameter `c_xy` controls the coupling strength. The implementation here also
+allows for tuning the parameters of each subsystem by introducing the constants
+`a₁`, `a₂`, `a₃`, `b₁`, `b₂`, `b₃`. Default values for these parameters are
+as in [1].
+
+## Equations of motion
+
+The dynamics is generated by the following vector field
+
+## Description
+
+```math
+\\begin{align*}
+\\dot x_1 &= -a_1(x_2 + x_3) \\\\
+\\dot x_2 &= a_2(x_1 + a_2x_2) \\\\
+\\dot x_3 &= a_1(a_2 + x_3(x_1 - a_3)) \\\\
+\\dot y_1 &= b_1(y_2 - y_1) \\\\
+\\dot y_2 &= y_1(b_2 - y_3) - y_2 +c_{xy}(x_2)^2 \\\\
+\\dot y_3 &= y_1 y_2 - b_3y_3
+\\end{align*}
+```
+
+with the coupling constant ``c_{xy} \\geq 0``.
+"""
+Base.@kwdef struct RosslerLorenzUnidir6{V, A1, A2, A3, B1, B2, B3, C} <: ContinuousDefinition
+    xi::V = [0.1, 0.2, 0.3, 0.05, 0.1, 0.15]
+    a₁::A1 = 6
+    a₂::A2 = 0.2
+    a₃::A3 = 5.7
+    b₁::B1 = 10
+    b₂::B2 = 28
+    b₃::B3 = 8/3
+    c_xy::C = 1.0
+end
+
+function system(definition::RosslerLorenzUnidir6)
+    return ContinuousDynamicalSystem(eom_rosslerlorenzunidir6, definition.xi, definition)
+end
+
+function eom_rosslerlorenzunidir6(u, p, t)
+    (; xi, a₁, a₂, a₃, b₁, b₂, b₃, c_xy) = p
+    x1, x2, x3, y1, y2, y3 = u
+
+    dx1 = -a₁*(x2 + x3)
+    dx2 = a₁*(x1 + a₂*x2)
+    dx3 = a₁*(a₂ + x3*(x1 - a₃))
+    dy1 = b₁*(-y1 + y2)
+    dy2 = b₂*y1 - y2 - y1*y3 + c_xy*(x2^2)
+    dy3 = y1*y2 - b₃*y3
+
+    return SVector{6}(dx1, dx2, dx3, dy1, dy2, dy3)
+end
