+++ src/example_systems/example_systems.jl
@@ -19,16 +19,15 @@ include("discrete/UlamLattice.jl")
 include("discrete/Var1.jl")
 include("discrete/Verdes3.jl")
 
-include("continuous/chuacircuits_driven.jl")
-include("continuous/chuacircuit_nscroll_sine.jl")
-include("continuous/hindmarsh_rose.jl")
-include("continuous/mediated_link.jl")
-include("continuous/lorenz_lorenz_bidir.jl")
-include("continuous/lorenz_lorenz_lorenz_bidir_forced.jl")
-include("continuous/lorenz_lorenz_lorenz_transitive.jl")
-include("continuous/repressilator.jl")
-include("continuous/rossler_rossler_bidir.jl")
-include("continuous/rossler_rossler_rossler_bidir_forced.jl")
-include("continuous/rossler_lorenz.jl")
-
-include("noise.jl")
+include("continuous/ChuaCircuitsBidir6.jl")
+include("continuous/ChuaScrollSine3.jl")
+include("continuous/HindmarshRose3.jl")
+include("continuous/MediatedLink9.jl")
+include("continuous/LorenzForced9.jl")
+include("continuous/LorenzTransitive9.jl")
+include("continuous/LorenzBidir6.jl")
+include("continuous/Repressilator6.jl")
+include("continuous/RosslerBidir6.jl")
+include("continuous/RosslerForced9.jl")
+include("continuous/RosslerLorenzUnidir6.jl")
+include("continuous/Thomas3.jl")
